GDA-5883 ScanDataPoint scan number is now an int

+++ uk.ac.gda.api/src/gda/scan/Scan.java
@@ -220,5 +220,5 @@ public interface Scan extends Serializable {
 	/**
 	 * @return The unique id of the scan. Null if not set
 	 */
-	public Long getScanNumber();
+	public int getScanNumber();
 }
